requirements.txt
aif360==0.4.0
alabaster @ file:///home/ktietz/src/ci/alabaster_1611921544520/work
anndata @ file:///home/conda/feedstock_root/build_artifacts/anndata_1637230564104/work
attrs==22.1.0
Babel @ file:///tmp/build/80754af9/babel_1620871417480/work
brotlipy==0.7.0
certifi==2021.10.8
cffi @ file:///opt/conda/conda-bld/cffi_1642701102775/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
cloudpickle==2.1.0
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
cryptography @ file:///tmp/build/80754af9/cryptography_1639414572950/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
docutils @ file:///tmp/build/80754af9/docutils_1620827980776/work
dunamai @ file:///home/conda/feedstock_root/build_artifacts/dunamai_1647341145295/work
easydict==1.9
faiss==1.7.2
fonttools==4.25.0
get_version @ file:///home/conda/feedstock_root/build_artifacts/get_version_1644239612226/work/get_version-3.5.4
h5py @ file:///tmp/build/80754af9/h5py_1637138488546/work
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
imagesize @ file:///tmp/build/80754af9/imagesize_1637939814114/work
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1638542770237/work
iniconfig==1.1.1
Jinja2 @ file:///tmp/build/80754af9/jinja2_1635780242639/work
joblib @ file:///tmp/build/80754af9/joblib_1635411271373/work
kiwisolver @ file:///opt/conda/conda-bld/kiwisolver_1638569886207/work
legacy-api-wrap==0.0.0
llvmlite==0.37.0
MarkupSafe @ file:///tmp/build/80754af9/markupsafe_1621523467000/work
matplotlib @ file:///tmp/build/80754af9/matplotlib-suite_1645455682260/work
memory-profiler==0.60.0
mkl-service==2.4.0
mock @ file:///tmp/build/80754af9/mock_1607622725907/work
munkres==1.1.4
natsort @ file:///tmp/build/80754af9/natsort_1611692802956/work
networkx @ file:///tmp/build/80754af9/networkx_1633639043937/work
numba @ file:///tmp/build/80754af9/numba_1635174338764/work
numexpr @ file:///tmp/build/80754af9/numexpr_1640689833592/work
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1622014603188/work
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.2.3
patsy==0.5.2
Pillow==9.0.1
pluggy==1.0.0
psutil==5.9.1
py==1.11.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
pynndescent @ file:///tmp/build/80754af9/pynndescent_1625654046303/work
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
pytest==7.1.2
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2021.3
PyYAML==6.0
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
scanpy @ file:///home/conda/feedstock_root/build_artifacts/scanpy_1636040642652/work
scikit-learn @ file:///tmp/build/80754af9/scikit-learn_1642617106979/work
scipy @ file:///home/conda/feedstock_root/build_artifacts/scipy_1604304774036/work
seaborn @ file:///tmp/build/80754af9/seaborn_1629307859561/work
shap==0.41.0
sinfo @ file:///home/conda/feedstock_root/build_artifacts/sinfo_1597764960136/work
sip==4.19.13
six @ file:///tmp/build/80754af9/six_1644875935023/work
slicer==0.0.7
snowballstemmer @ file:///tmp/build/80754af9/snowballstemmer_1637937080595/work
Sphinx @ file:///opt/conda/conda-bld/sphinx_1643644169832/work
sphinxcontrib-applehelp @ file:///home/ktietz/src/ci/sphinxcontrib-applehelp_1611920841464/work
sphinxcontrib-devhelp @ file:///home/ktietz/src/ci/sphinxcontrib-devhelp_1611920923094/work
sphinxcontrib-htmlhelp @ file:///tmp/build/80754af9/sphinxcontrib-htmlhelp_1623945626792/work
sphinxcontrib-jsmath @ file:///home/ktietz/src/ci/sphinxcontrib-jsmath_1611920942228/work
sphinxcontrib-qthelp @ file:///home/ktietz/src/ci/sphinxcontrib-qthelp_1611921055322/work
sphinxcontrib-serializinghtml @ file:///tmp/build/80754af9/sphinxcontrib-serializinghtml_1624451540180/work
statsmodels @ file:///tmp/build/80754af9/statsmodels_1614022848006/work
stdlib-list @ file:///home/conda/feedstock_root/build_artifacts/stdlib-list_1602639452997/work
tables @ file:///tmp/build/80754af9/pytables_1607975397488/work
tempeh==0.1.12
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
tomli==2.0.1
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
tornado @ file:///tmp/build/80754af9/tornado_1606942317143/work
tqdm @ file:///opt/conda/conda-bld/tqdm_1647339053476/work
tsnecuda==3.0.0+cu110
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
umap-learn @ file:///home/conda/feedstock_root/build_artifacts/umap-learn_1642213314285/work
urllib3 @ file:///opt/conda/conda-bld/urllib3_1643638302206/work
zipp @ file:///opt/conda/conda-bld/zipp_1641824620731/work
